Server now serves files and lists directories via tcp.
All files/directories in the 'files' directory will be exposed.
The user will connect via tcp, and can list a directory by typing a relative
path to that directory, or get a file by giving a path to that file.
I.E to list files in the root 'files' directory, the user would send '/'